Aula dia 05/05 e checkpoint atualizado

+++ 06-exercicios/src/br/com/fiap/exercicio01/view/Exercicio.java
@@ -34,7 +34,7 @@ public class Exercicio {
 		
 		//Ler o comprimento do barco
 		System.out.println("Digite o comprimento: ");
-		int comprimento = sc.nextInt();
+		double comprimento = sc.nextDouble();
 		barco.setComprimento(comprimento);
 		
 		//Ler o ano de fabricação
